// frontend/src/api/generated/model/pullRequestDetail.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * OpenCode Studio API
 * API for OpenCode Studio - AI-powered development platform
 * OpenAPI spec version: 0.1.0
 */
import type { PullRequestDetailBody } from './pullRequestDetailBody';
import type { PullRequestDetailCiStatus } from './pullRequestDetailCiStatus';
import type { Label } from './label';
import type { PullRequestDetailMergeable } from './pullRequestDetailMergeable';
import type { PullRequestDetailMergeableState } from './pullRequestDetailMergeableState';
import type { PullRequestDetailMergedAt } from './pullRequestDetailMergedAt';
import type { GitHubUser } from './gitHubUser';
import type { PrState } from './prState';

export interface PullRequestDetail {
  /** @minimum 0 */
  additions: number;
  base_branch: string;
  body?: PullRequestDetailBody;
  /** @minimum 0 */
  changed_files: number;
  ci_status?: PullRequestDetailCiStatus;
  /** @minimum 0 */
  comments_count: number;
  created_at: string;
  /** @minimum 0 */
  deletions: number;
  draft: boolean;
  head_branch: string;
  html_url: string;
  labels: Label[];
  mergeable?: PullRequestDetailMergeable;
  mergeable_state?: PullRequestDetailMergeableState;
  merged_at?: PullRequestDetailMergedAt;
  /** @minimum 0 */
  number: number;
  requested_reviewers: GitHubUser[];
  /** @minimum 0 */
  review_comments_count: number;
  state: PrState;
  title: string;
  updated_at: string;
  user: GitHubUser;
}